requirements.txt
docker>=6.1.3
pyautogen>=0.1.10